+++ widgets/TabQuery.py
@@ -66,6 +66,15 @@ class TabQuery(BorderBox):
 		self.__inOut.setTypeList(self.__db.getTypeList())
 	
 	def __handlerDelete(self, widget, data=None):
+		messageText = config.getDisplayString('TEDiaDeleteWarning')
+		dialog = Gtk.MessageDialog(message_type=Gtk.MessageType.WARNING, buttons=Gtk.ButtonsType.YES_NO, message_format=messageText)
+
+		response = dialog.run()
+		dialog.destroy()
+
+		if response != Gtk.ResponseType.YES:
+			return
+
 		self.__db.deleteVoc(self.__nextVoc)
 		self.__getNextVoc()
 		self.__query()
